requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.31
alembic==1.12.1
# psycopg2-binary==2.9.9  # PostgreSQL driver - for production use later
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
redis==5.0.1
email-validator==2.1.0
pydantic-settings==2.1.0
# pydantic and pydantic-settings will be installed by FastAPI automatically 